Add stages within TeaLeaf to PETSc logging.
Split TeaLeaf into Assemble and Solve stages (plus everything else, which is classed as "Main Stage" by PETSc. This allows us to identify whether some parts of the code are threaded more than others. Note that PETSc delegated responsibility for threading matrix assembly to the application code and this is not currently threaded in TeaLeaf.