Added computation of ZTA. Attempted to make the code compatible with PETSc 3.4.2, but this failed due to the lack of a fortran interface for DMSetMatType.
Code segfaults with PETSc 3.6.2 when using -mat_view in the initial construction of ZTA during the symbolic factorisation. I checked whether this was due to not having assembled A at this stage, but even with A being populated the code crashes in the relocated symbolic factorisation. This must be a logic error in PETSc i.e. trying to print matrix entries prior to the storage for them being allocated.